+++ Domain/RepositoryInterfaces/IVehicleInterfaces.cs
@@ -5,10 +5,10 @@ namespace Domain.RepositoryInterfaces
     public interface IVehicleRepository
     {
         Task<IEnumerable<Vehicle>> GetAllAsync();
-        Task<Vehicle?> GetByIdAsync(double id);
+        Task<Vehicle?> GetByIdAsync(long id);
         Task<Vehicle?> GetByLicensePlateAsync(string plate);
         Task AddAsync(Vehicle vehicle);
         Task UpdateAsync(Vehicle vehicle);
-        Task DeleteAsync(double id);
+        Task DeleteAsync(long id);
     }
 }
